Initial V1.0 of the program
Version not to be used on the vehicle - bench tested only. Contains functionality for automatic height adjustment using an input signal from 1 potentiometer (vehicle height sensor), as well as mode selection using a button, and interaction with O/S/R solenoid, N/S/R solenoid, vent solenoid, and the relay for the pump.