Update LR_D2_SLS_CTRL.ino
Not for use on the vehicle - only bench tested. Removed automatic height adjustment during drive. When system is initially run it will attempt to level suspension to the "Normal" mode, and then deactivate. It will only activate again when the mode button is selected, at which point it will attempt to level to the selected mode, and then deactivate again. This potential improvement was made to avoid unnecessary activation of the system during driving conditions with increased suspension travel.